+++ client-spark/spark2/src/main/java/org/apache/spark/shuffle/writer/RssShuffleWriter.java
@@ -184,7 +184,7 @@ public class RssShuffleWriter<K, V, C> extends ShuffleWriter<K, V> {
       String appId,
       int shuffleId,
       String taskId,
-      int taskAttemptId,
+      long taskAttemptId,
       ShuffleWriteMetrics shuffleWriteMetrics,
       RssShuffleManager shuffleManager,
       SparkConf sparkConf,
