+++ client-tez/src/main/java/org/apache/tez/runtime/library/output/RssUnorderedKVOutput.java
@@ -222,7 +222,7 @@ public class RssUnorderedKVOutput extends AbstractLogicalOutput {
   public void start() throws Exception {
     if (!isStarted.get()) {
       memoryUpdateCallbackHandler.validateUpdateReceived();
-      int rssTaskAttemptId = RssTezUtils.createRssTaskAttemptId(taskAttemptId, conf);
+      long rssTaskAttemptId = RssTezUtils.createRssTaskAttemptId(taskAttemptId, conf);
       sorter =
           new RssUnSorter(
               taskAttemptId,
