+++ client-tez/src/main/java/org/apache/tez/runtime/library/output/RssUnorderedPartitionedKVOutput.java
@@ -220,7 +220,7 @@ public class RssUnorderedPartitionedKVOutput extends AbstractLogicalOutput {
   public void start() throws Exception {
     if (!isStarted.get()) {
       memoryUpdateCallbackHandler.validateUpdateReceived();
-      int rssTaskAttemptId = RssTezUtils.createRssTaskAttemptId(taskAttemptId, conf);
+      long rssTaskAttemptId = RssTezUtils.createRssTaskAttemptId(taskAttemptId, conf);
       sorter =
           new RssUnSorter(
               taskAttemptId,
