+++ client-tez/src/test/java/org/apache/tez/runtime/library/common/sort/impl/RssUnSorterTest.java
@@ -83,7 +83,7 @@ public class RssUnSorterTest {
 
     long initialMemoryAvailable = 10240000;
     int shuffleId = 1001;
-    int rssTaskAttemptId = RssTezUtils.createRssTaskAttemptId(tezTaskAttemptID, 3);
+    long rssTaskAttemptId = RssTezUtils.createRssTaskAttemptId(tezTaskAttemptID, 3);
 
     RssUnSorter rssSorter =
         new RssUnSorter(
