+++ client/src/test/java/org/apache/uniffle/client/impl/ShuffleReadClientImplTest.java
@@ -693,7 +693,7 @@ public class ShuffleReadClientImplTest extends HadoopTestBase {
       int num,
       int length,
       int partitionId,
-      int taskAttemptId,
+      long taskAttemptId,
       Map<Long, byte[]> expectedData,
       Roaring64NavigableMap blockIdBitmap,
       BlockIdLayout layout)
@@ -717,7 +717,7 @@ public class ShuffleReadClientImplTest extends HadoopTestBase {
       int num,
       int length,
       int partitionId,
-      int taskAttemptId,
+      long taskAttemptId,
       Map<Long, byte[]> expectedData,
       Roaring64NavigableMap blockIdBitmap)
       throws Exception {
@@ -737,7 +737,7 @@ public class ShuffleReadClientImplTest extends HadoopTestBase {
       int num,
       int length,
       int partitionId,
-      int taskAttemptId,
+      long taskAttemptId,
       Map<Long, byte[]> expectedData,
       Roaring64NavigableMap blockIdBitmap)
       throws Exception {
